HuongGiang : update type gia and tong chi phi Double to Bigdecimal

+++ src/main/java/org/example/wms_be/data/response/inbound/PurchaseRequestDetailsIbResp.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -15,9 +17,9 @@ public class PurchaseRequestDetailsIbResp {
     private String  maLoHang;
     private String  tenKhachHang;
     private String  tenSanPham;
-    private Double  soLuong;
-    private Double  gia;
-    private Double  tongChiPhi;
+    private Double soLuong;
+    private BigDecimal   gia;
+    private BigDecimal   tongChiPhi;
     private String  ngayNhapDuKien;
     private Integer sysIdKhachHang;
     private Integer sysIdSanPham;
